add top projects,search project

+++ CrowdFundCore/Options/FundingOption.cs
@@ -8,7 +8,7 @@ namespace CrowdFundCore.Options
     public class FundingOption
     {
         public int Id { get; set; }
-        public int Ammount { get; set; }
+        public decimal Ammount { get; set; }
     }
 }
